+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+Release 5.4.1:
+ - Fix encoding `dcql_query` in authentication request, it is now a string
+ - Provide default values for RQES data classes
+
 Release 5.4.0:
 - Extend support for POTENTIAL UC5: Remote qualified electronic signatures
   - Update data classes in `rqes-data-classes`